Open URL in eclipse (internal or external) instead of chat browser.
Clicking any links in the chat will result in the link being opened in either an internal or external browser based on the preference the user has set in eclipse.

This is detected in the browser's location handler in ChatBrowser.java.

Removed anchor.target = "_blank" since it's not needed and the browser objects event handler will process the navigation.